Trying to add different colors to different strips, Many overloaded functions are added to take care of translation in Z direction which is required for different RPCs and different Scintillator Plane,
But the problem remains is that, still not able to display multiple RPCs,
Able to display on On Plane at a time.

Because of these new changes the commit is now switched to new Branch raman/Tracking_v2.
Once it starts working it will be merged to raman/Tracking_v1